Set argdatalen to the right value.
It turned out we set it to some garbage value. This coincidentally made
the unit testing binary work a bit, but this is obviously not the path
forward.

Use the value of argc to recompute arg_end. We can then use that field
to properly derive the size of the argument data buffer.