Work towards having a properly set st_filetype.
Add a function, cloudabi_convert_filetype(), that we can use to inspect
the file descriptor and extract its type accurately. For now, already
extend it to properly recognize sockets. The _simple() function can now
be used to look at st_mode.

For fstatat(), we use _simple(). For fstat() we can use the regular
function.