Just use a semaphore to lock down the futex pool.
It turns out that if we use a spinlock for the futexes, we also cannot
enable page faulting while doing loads, stores and compare-and-exchanges.
Instead of making the entire code spinlock proof, let's just fall back
to using a slow semaphore. It also has the advantage that we can stick
to GFP_KERNEL for now.