The large system call prototype change.
Let's go ahead and let all system call implementations use exactly the
same prototypes as in userspace. This makes them a lot more friendly to
implement and also simplifies supporting both 32 and 64 bit processes in
a single kernel.

We now put all the nastiness of extracting the parameters from the right
locations in the registers in cloudabi64_syscalls.c. This generated code
now maps structures over both the input and output and invokes the
system call implementations.